shifted alerts to another tab
