+++ src/components/VisualisationComponent.tsx
@@ -2,8 +2,6 @@ import React, { useEffect, useState } from "react";
 import { Responsive, WidthProvider } from "react-grid-layout";
 import "react-grid-layout/css/styles.css";
 import "react-resizable/css/styles.css";
-import { Box, Grid, Tab, Tabs, Typography, styled } from "@mui/material";
-import { DataGrid, GridRowModel } from "@mui/x-data-grid";
 import { Patient } from "@/models/patient";
 import { updatePatientLayoutByPatientId } from "@/pages/api/patients_api";
 import { Layouts } from "react-grid-layout";
@@ -18,123 +16,11 @@ import LineChartComponent from "@/components/LineChart";
 
 const ResponsiveGridLayout = WidthProvider(Responsive);
 
-const alertColumns = [
-  { field: "id", headerName: "ID" },
-  { field: "status", headerName: "Status" },
-  { field: "description", headerName: "Description" },
-  { field: "notes", headerName: "Notes" },
-];
-
-const StyledDataGrid = styled(DataGrid)(({ theme }) => ({
-  "& .alert-open": {
-    backgroundColor: "pink",
-    "&:hover": {
-      cursor: "pointer",
-      backgroundColor: "pink",
-    },
-  },
-  "& .alert-handling": {
-    backgroundColor: "#FFA829",
-    "&:hover": {
-      cursor: "pointer",
-      backgroundColor: "#FFA829",
-    },
-  },
-  "& .alert-complete": {
-    backgroundColor: "lightgreen",
-    "&:hover": {
-      cursor: "pointer",
-      backgroundColor: "lightgreen",
-    },
-  },
-}));
-
-interface TabPanelProps {
-  status: string;
-  index: number;
-  value: number;
-}
-
 interface ComponentProp {
   patient: Patient | undefined;
 }
 
-function a11yProps(index: number) {
-  return {
-    id: `simple-tab-${index}`,
-    "aria-controls": `simple-tabpanel-${index}`,
-  };
-}
-
 export default function VisualisationComponent(prop: ComponentProp) {
-  const [value, setValue] = useState(0);
-
-  const handleChange = (event: React.SyntheticEvent, newValue: number) => {
-    setValue(newValue);
-  };
-
-  function getAlerts(status: string) {
-    const listOfAlerts: GridRowModel[] = [];
-    console.log(patientAlerts);
-    if (patientAlerts !== undefined) {
-      for (let i = 0; i < patientAlerts.length; i++) {
-        console.log(patientAlerts[i]);
-        if (patientAlerts[i].status == status) {
-          listOfAlerts.push({
-            status: patientAlerts[i].status,
-            description: patientAlerts[i].description,
-            notes: patientAlerts[i].notes,
-          });
-        }
-      }
-    }
-    console.log(listOfAlerts);
-    return listOfAlerts.map((alert, index) => ({
-      id: index + 1,
-      ...alert,
-    }));
-  }
-
-  function CustomTabPanel(props: TabPanelProps) {
-    const { status, value, index, ...other } = props;
-
-    return (
-      <Box>
-        {value === index && (
-          <Box>
-            <Grid item xs={6} style={{ flex: 1 }}>
-              {prop.patient?.alerts?.length != undefined &&
-              getAlerts(status).length > 0 ? (
-                <Box>
-                  <StyledDataGrid
-                    aria-label="Alerts"
-                    columns={alertColumns}
-                    rows={getAlerts(status)}
-                    autoHeight
-                    rowHeight={100}
-                    getRowClassName={(params) => `alert-${params.row.status}`}
-                    sx={{
-                      "& .MuiDataGrid-cellContent": {
-                        whiteSpace: "normal !important",
-                        wordWrap: "break-word !important",
-                      },
-                    }}
-                  />
-                </Box>
-              ) : (
-                <div>
-                  <Typography marginTop={2} variant="body1">
-                    No {status} alerts
-                  </Typography>
-                </div>
-              )}
-            </Grid>
-          </Box>
-        )}
-      </Box>
-    );
-  }
-
   const [retrieveLayout, setLayouts] = useState(prop.patient?.layout);
 
   const onLayoutChange = (_: Layout[], allLayouts: Layouts) => {
@@ -163,16 +49,12 @@ export default function VisualisationComponent(prop: ComponentProp) {
   }, [prop.patient?.layout]);
 
   const [patientVitals, setPatientVitals] = useState<Vital>();
-  const [patientAlerts, setPatientAlerts] = useState<Alert[]>();
 
   useEffect(() => {
     fetchVitalByVitalId(prop?.patient?.vital).then((res) =>
       setPatientVitals(res)
     );
-    fetchAlertsByPatientId(prop?.patient?._id).then((res) =>
-      setPatientAlerts(res)
-    );
-  }, [prop?.patient?.vital, prop?.patient?._id]);
+  }, [prop?.patient?.vital]);
 
   useEffect(() => {
     if (patientVitals) {
@@ -228,9 +110,6 @@ export default function VisualisationComponent(prop: ComponentProp) {
     bloodPressureSys?: number;
     bloodPressureDia?: number;
     spO2?: number;
-    // bloodPressureSys?: string;
-    // bloodPressureDia?: string;
-    // spO2?: string;
   }
 
   const router = useRouter();
@@ -339,31 +218,6 @@ export default function VisualisationComponent(prop: ComponentProp) {
       </div>
       <div key="tp">
         <LineChartComponent data={tempData!} vital="tp" />
-      </div>
-      <div key="alerts">
-        <Box sx={{ borderBottom: 1, borderColor: "divider" }}>
-          <h3>Alerts</h3>
-          <Tabs
-            value={value}
-            onChange={handleChange}
-            aria-label="basic tabs example"
-          >
-            <Tab label="Open" {...a11yProps(0)} />
-            <Tab label="Handling" {...a11yProps(1)} />
-            <Tab label="Completed" {...a11yProps(2)} />
-          </Tabs>
-        </Box>
-        <CustomTabPanel value={value} index={0} status="open"></CustomTabPanel>
-        <CustomTabPanel
-          value={value}
-          index={1}
-          status="handling"
-        ></CustomTabPanel>
-        <CustomTabPanel
-          value={value}
-          index={2}
-          status="complete"
-        ></CustomTabPanel>
       </div>
     </ResponsiveGridLayout>
   );
